feat: mvp

+++ src/modules/group/group.service.ts
@@ -1,12 +1,18 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
+import { ErrorMap } from 'src/shared/response/error.map';
+import { User } from '../user/entities/user.entity';
+import { UserPolicyService } from '../user/user-policy.service';
 import { CreateGroupDto } from './dto/create-group.dto';
 import { UpdateGroupDto } from './dto/update-group.dto';
-import { GroupRepository } from './group.repository';
 import { Group } from './entities/group.entity';
+import { GroupRepository } from './group.repository';
 
 @Injectable()
 export class GroupService {
-  constructor(private readonly groupRepository: GroupRepository) { }
+  constructor(
+    private readonly policy: UserPolicyService,
+    private readonly groupRepository: GroupRepository,
+  ) {}
 
   create(createGroupDto: CreateGroupDto): Promise<Group> {
     return this.groupRepository.createGroup(createGroupDto);
@@ -24,11 +30,17 @@ export class GroupService {
     return this.groupRepository.findGroup(id);
   }
 
-  update(id: string, updateGroupDto: UpdateGroupDto): Promise<boolean> {
+  update(id: string, updateGroupDto: UpdateGroupDto, callerUser: User): Promise<boolean> {
+    if (!this.policy.isManagerHasPermissionByGroupId(id, callerUser)) {
+      throw new ForbiddenException(ErrorMap.FORBIDDEN_ERROR);
+    }
     return this.groupRepository.updateGroup(id, updateGroupDto);
   }
 
-  remove(id: string): Promise<boolean> {
+  remove(id: string, callerUser: User): Promise<boolean> {
+    if (!this.policy.isManagerHasPermissionByGroupId(id, callerUser)) {
+      throw new ForbiddenException(ErrorMap.FORBIDDEN_ERROR);
+    }
     return this.groupRepository.removeGroup(id);
   }
 }
